requirements.txt
# Core Data & ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0  # For GNN
transformers>=4.30.0  # For BERT/DistilBERT embeddings

# NLP & Text Processing
nltk>=3.8
spacy>=3.6.0
sentence-transformers>=2.2.0  # For embeddings

# Financial Data & APIs
yfinance>=0.2.0
fredapi>=0.5.0  # Federal Reserve Economic Data
requests>=2.31.0

# Graph Processing
networkx>=3.1
dgl>=1.0.0  # Deep Graph Library for GNN

# Training & Optimization
optuna>=3.0.0  # Hyperparameter optimization
wandb>=0.15.0  # Experiment tracking

# Evaluation & Metrics
scikit-optimize>=0.9.0
numpy-financial>=1.0.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter & Interactive
jupyter>=1.0.0
ipython>=8.14.0
